matlab
Former-commit-id: 6e3de377989272fec53ab1d55653c7c0302ccbe9 [formerly ca4c1f1e03f0f02e3cdf9582e931feecafd42ce4] [formerly 5fa8ae883a96db774ece88e3bb5ab89d84b98131 [formerly d7b1e7b3a7664ff7568368a977427a7e74741e23 [formerly 809f1124c52598f3fcddc66cb46106c848e4a374]]]
Former-commit-id: 6e3de377989272fec53ab1d55653c7c0302ccbe9 [formerly ca4c1f1e03f0f02e3cdf9582e931feecafd42ce4]
Former-commit-id: ea5812ddc63efd81eb016e7e48f9c59a9c510123 [formerly cc3e484b7d4938e4d315c7043bbcb3a24e51db45 [formerly 9f9d2984698ac0d0410ff3fdb14a5c07cecf717f]]
Former-commit-id: 1df8ade847f7b0d6115bfe7c70259599bb6fa241
Former-commit-id: ebb67de94a9a5b7cce6994e4d60fabcd04c406a6

+++ matlab/NIfTI_toolbox/save_nii_v2.m
@@ -48,7 +48,7 @@
 %  - Jimmy Shen (jimmy@rotman-baycrest.on.ca)
 %  - "old_RGB" related codes in "save_nii.m" are added by Mike Harms (2006.06.28) 
 %
-function save_nii_v2(nii, fileprefix, old_nii_fname,datatype)
+function save_nii_v2(nii, fileprefix, old_nii_fname,datatype,reorient)
    dbstop if error
    if ~exist('nii','var') | isempty(nii) %| ~isfield(nii,'hdr') | ...
 	%~isfield(nii,'img') | ~exist('fileprefix','var') | isempty(fileprefix)
@@ -142,10 +142,12 @@ function save_nii_v2(nii, fileprefix, old_nii_fname,datatype)
    write_nii(nii, filetype, fileprefix, old_RGB);
    % reorient in RPI
    
-   if exist('old_nii_fname','var')
+   if ~exist('reorient','var'), reorient=1; end
+   if exist('old_nii_fname','var') && reorient
        [~,orient_dest] = unix(['sct_image -i ' old_nii_fname ' -getorient']); orient_dest=orient_dest(1:end-1);
        unix(['sct_image -i ' fileprefix '.nii -setorient ' orient_dest]);
-       unix(['mv ' fileprefix '_RPI.nii ' fileprefix '.nii'])
+       if exist([fileprefix '.nii.gz'],'file'), unix(['rm ' fileprefix '.nii.gz']); end
+       unix(['mv ' fileprefix '_RPI.nii ' fileprefix '.nii']);
        unix(['fslcpgeom ' old_nii_fname ' ' fileprefix '.nii -d']);
    end
